refact: distribute cards into modules
