+++ app/src/main/java/com/github/felipecastilhos/pokedexandroid/features/pokemon/PokemonSearchScreen.kt
@@ -1,13 +1,10 @@
 package com.github.felipecastilhos.pokedexandroid.features.pokemon
 
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.lazy.grid.GridCells
 import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
 import androidx.compose.foundation.lazy.grid.items
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
@@ -16,7 +13,7 @@ import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.github.felipecastilhos.pokedexandroid.features.pokemon.domain.viewmodel.PokemonListEntryUiData
 import com.github.felipecastilhos.pokedexandroid.features.pokemon.domain.viewmodel.PokemonListViewModel
-import com.github.felipecastilhos.pokedexandroid.uicatalog.components.PokemonListCard
+import com.github.felipecastilhos.pokedexandroid.features.pokemon.presentation.components.PokemonListCard
 
 @Composable
 fun PokemonSearchScreen(
